updated newton-based IL sim2sim transfer
