+++ scripts/sim2sim_transfer/config/physx_to_newton_g1.yaml
