+++ scripts/sim2sim_transfer/rsl_rl_transfer.py
@@ -13,7 +13,6 @@ from isaaclab.app import AppLauncher
 
 # local imports
 from scripts.reinforcement_learning.rsl_rl import cli_args  # isort: skip
-from scripts.newton_sim2sim.policy_mapping import get_joint_mappings  # isort: skip
 
 # add argparse arguments
 parser = argparse.ArgumentParser(description="Train an RL agent with RSL-RL.")
@@ -57,6 +56,7 @@ import gymnasium as gym
 import os
 import time
 import torch
+import yaml
 
 from rsl_rl.runners import OnPolicyRunner
 
@@ -75,6 +75,67 @@ import isaaclab_tasks  # noqa: F401
 from isaaclab_tasks.utils import get_checkpoint_path, parse_env_cfg
 
 # PLACEHOLDER: Extension template (do not remove this comment)
+
+
+def get_joint_mappings(args_cli, action_space_dim):
+    """Get joint mappings based on command line arguments.
+
+    Args:
+            args_cli: Command line arguments
+            action_space_dim: Dimension of the action space (number of joints)
+
+    Returns:
+            tuple: (source_to_target_list, target_to_source_list, source_to_target_obs_list)
+    """
+    num_joints = action_space_dim
+    if args_cli.policy_transfer_file:
+        # Load from YAML file
+        try:
+            with open(args_cli.policy_transfer_file) as file:
+                config = yaml.safe_load(file)
+        except Exception as e:
+            raise RuntimeError(f"Failed to load joint mapping from {args_cli.policy_transfer_file}: {e}")
+
+        source_joint_names = config["source_joint_names"]
+        target_joint_names = config["target_joint_names"]
+        # Find joint mapping
+        source_to_target = []
+        target_to_source = []
+
+        # Create source to target mapping
+        for joint_name in source_joint_names:
+            if joint_name in target_joint_names:
+                source_to_target.append(target_joint_names.index(joint_name))
+            else:
+                raise ValueError(f"Joint '{joint_name}' not found in target joint names")
+
+        # Create target to source mapping
+        for joint_name in target_joint_names:
+            if joint_name in source_joint_names:
+                target_to_source.append(source_joint_names.index(joint_name))
+            else:
+                raise ValueError(f"Joint '{joint_name}' not found in source joint names")
+        print(f"[INFO] Loaded joint mapping for policy transfer from YAML: {args_cli.policy_transfer_file}")
+        assert (
+            len(source_to_target) == len(target_to_source) == num_joints
+        ), "Number of source and target joints must match"
+    else:
+        # Use identity mapping (one-to-one)
+        identity_map = list(range(num_joints))
+        source_to_target, target_to_source = identity_map, identity_map
+
+    # Create observation mapping (first 12 values stay the same for locomotion examples, then map joint-related values)
+    obs_map = (
+        [0, 1, 2]
+        + [3, 4, 5]
+        + [6, 7, 8]
+        + [9, 10, 11]
+        + [i + 12 + num_joints * 0 for i in source_to_target]
+        + [i + 12 + num_joints * 1 for i in source_to_target]
+        + [i + 12 + num_joints * 2 for i in source_to_target]
+    )
+
+    return source_to_target, target_to_source, obs_map
 
 
 def main():
@@ -141,17 +202,23 @@ def main():
         # version 2.2 and below
         policy_nn = ppo_runner.alg.actor_critic
 
+    # extract the normalizer
+    if hasattr(policy_nn, "actor_obs_normalizer"):
+        normalizer = policy_nn.actor_obs_normalizer
+    elif hasattr(policy_nn, "student_obs_normalizer"):
+        normalizer = policy_nn.student_obs_normalizer
+    else:
+        normalizer = None
+
     # export policy to onnx/jit
     export_model_dir = os.path.join(os.path.dirname(resume_path), "exported")
-    export_policy_as_jit(policy_nn, ppo_runner.obs_normalizer, path=export_model_dir, filename="policy.pt")
-    export_policy_as_onnx(
-        policy_nn, normalizer=ppo_runner.obs_normalizer, path=export_model_dir, filename="policy.onnx"
-    )
+    export_policy_as_jit(policy_nn, normalizer=normalizer, path=export_model_dir, filename="policy.pt")
+    export_policy_as_onnx(policy_nn, normalizer=normalizer, path=export_model_dir, filename="policy.onnx")
 
     dt = env.unwrapped.step_dt
 
     # reset environment
-    obs, _ = env.get_observations()
+    obs = env.get_observations()
     timestep = 0
     # Get joint mappings for policy transfer
     _, target_to_source, obs_map = get_joint_mappings(args_cli, env.action_space.shape[1])
